refactor: header와 관련된 파일을 분류하는 header 디렉터리 생성
