+++ src/step2-index.js
@@ -1,5 +1,5 @@
 import createGameBox from "./view/web/layers/createGameBox.js";
-import createHeader from "./view/web/layers/header.js";
+import createHeader from "./view/web/layers/header/header.js";
 import readLottoPriceInput from "./view/web/modules/readLottoPriceInput.js";
 
 const app = async () => {
